refactor: simplify app structure by removing tab navigation
- Remove tab-based navigation structure ((tabs) directory)
- Consolidate all functionality into single screen at app/index.tsx
- Update root layout to reference index.tsx directly
- Preserve all existing search, address selection, and waste collection features
- Maintain streamlined UI design without tabs for better user experience
