--- app/(tabs)/_layout.tsx
@@ -1,48 +1,0 @@
-// External package imports
-import { Tabs } from "expo-router";
-import { Platform } from "react-native";
-
-// Internal absolute imports
-import { HapticTab } from "@/components/HapticTab";
-import { IconSymbol } from "@/components/ui/IconSymbol";
-import TabBarBackground from "@/components/ui/TabBarBackground";
-import { Colors } from "@/constants/Colors";
-import { useColorScheme } from "@/hooks/useColorScheme";
-
-export default function TabLayout() {
-	const colorScheme = useColorScheme();
-
-	return (
-		<Tabs
-			screenOptions={{
-				tabBarActiveTintColor: Colors[colorScheme ?? "light"].tint,
-				headerShown: false,
-				tabBarButton: HapticTab,
-				tabBarBackground: TabBarBackground,
-				tabBarStyle: Platform.select({
-					ios: {
-						// Use a transparent background on iOS to show the blur effect
-						position: "absolute",
-					},
-					default: {},
-				}),
-			}}
-		>
-			<Tabs.Screen
-				name="index"
-				options={{
-					href: null,
-				}}
-			/>
-			<Tabs.Screen
-				name="search"
-				options={{
-					title: "Search",
-					tabBarIcon: ({ color }) => (
-						<IconSymbol size={28} name="magnifyingglass" color={color} />
-					),
-				}}
-			/>
-		</Tabs>
-	);
-}
