--- app/(tabs)/index.tsx
@@ -1,5 +1,0 @@
-import { Redirect } from "expo-router";
-
-export default function TabIndex() {
-	return <Redirect href="/search" />;
-}
