+++ app/_layout.tsx
@@ -32,7 +32,7 @@ export default function RootLayout() {
 		<ConvexProvider client={convex}>
 			<ThemeProvider value={colorScheme === "dark" ? DarkTheme : DefaultTheme}>
 				<Stack>
-					<Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+					<Stack.Screen name="index" options={{ headerShown: false }} />
 					<Stack.Screen name="+not-found" />
 				</Stack>
 				<StatusBar style="auto" />
